Upgrade dependencies (Next.js 16 / TypeScript 6 / ESLint 9 flat config)
主要パッケージを最新へアップグレードし、それに伴う破壊的変更を修正。

## バージョンアップ
- next / eslint-config-next: 15 → 16.2.7
- typescript: 5 → 6.0.3
- @types/node: 24 → 25
- react / react-dom: 19.2.7、highlight.js 11.11.1、microcms-js-sdk 3.4.0、prettier 3.8.3

## 破壊的変更への対応
- tsconfig: target es5 → es2017（TS6 で es5 は TS5107 エラー）
- middleware.ts → proxy.ts にリネーム（Next.js 16 で middleware は非推奨、関数名も proxy）
- ESLint: next lint コマンドが Next.js 16 で削除されたため lint スクリプトを `eslint .` に変更
- ESLint: eslint-config-next@16 は flat config 専用のため .eslintrc.json を eslint.config.mjs に移行
- ESLint: 10.x は eslint-plugin-react が非互換でクラッシュするため 9.39.4 に固定
- postcss: next@16 が脆弱な 8.4.31 を要求するため overrides で 8.5.x に固定（脆弱性 0 件）

## クリーンアップ
- next.config の CDN-Cache-Control から _next/static 等の静的アセットを除外
- キャッシュ TTL を constants/cache.ts に集約（proxy.ts / next.config.ts が import）
- ページの revalidate はリテラル必須（Next.js 静的解析の制約）のためガードコメント追加

## 検証
- tsc --noEmit / npm run lint / npm run build / npm audit すべて成功（脆弱性 0 件）
