--- .eslintrc.json
@@ -1,3 +1,0 @@
-{
-  "extends": ["next", "next/core-web-vitals", "prettier"]
-}
