+++ app/search/p/[current]/page.tsx
@@ -8,6 +8,7 @@ type Props = {
   searchParams: Promise<{ q?: string }>;
 };
 
+// Next.js が静的解析するためリテラル必須（constants/cache.ts: REVALIDATE_SECONDS と同期）
 export const revalidate = 60;
 
 export default async function Page(props: Props) {
