+++ app/tags/[tagId]/p/[current]/page.tsx
@@ -7,6 +7,7 @@ type Props = {
   params: Promise<{ tagId: string; current: string }>;
 };
 
+// Next.js が静的解析するためリテラル必須（constants/cache.ts: REVALIDATE_SECONDS と同期）
 export const revalidate = 60;
 
 export default async function Page(props: Props) {
