+++ eslint.config.mjs
@@ -1,0 +1,6 @@
+import nextCoreWebVitals from 'eslint-config-next/core-web-vitals';
+import prettier from 'eslint-config-prettier/flat';
+
+const config = [...nextCoreWebVitals, prettier];
+
+export default config;
