+++ package.json
@@ -6,7 +6,7 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint",
+    "lint": "eslint .",
     "format": "prettier --write --ignore-path .gitignore './**/*.{js,jsx,ts,tsx,json,css}'"
   },
   "prettier": {
@@ -20,22 +20,25 @@
   "engines": {
     "node": ">=22.x"
   },
+  "overrides": {
+    "postcss": "^8.5.15"
+  },
   "dependencies": {
-    "@types/node": "^24.1.0",
-    "@types/react": "^19.2.10",
+    "@types/node": "^25.9.2",
+    "@types/react": "^19.2.17",
     "@types/react-dom": "^19.2.3",
     "cheerio": "^1.0.0",
     "date-fns-tz": "^3.2.0",
     "encoding": "^0.1.13",
-    "eslint": "^9.23.0",
-    "eslint-config-next": "^15.2.3",
-    "eslint-config-prettier": "^10.1.1",
-    "highlight.js": "^11.7.0",
-    "microcms-js-sdk": "^3.2.0",
-    "next": "^15.5.10",
-    "prettier": "^3.5.3",
-    "react": "^19.2.4",
-    "react-dom": "^19.2.4",
-    "typescript": "^5.8.2"
+    "eslint": "^9.39.4",
+    "eslint-config-next": "^16.2.7",
+    "eslint-config-prettier": "^10.1.8",
+    "highlight.js": "^11.11.1",
+    "microcms-js-sdk": "^3.4.0",
+    "next": "^16.2.7",
+    "prettier": "^3.8.3",
+    "react": "^19.2.7",
+    "react-dom": "^19.2.7",
+    "typescript": "^6.0.3"
   }
 }
